Splitting some C# related extensions in a separate file
