+++ src/Stryker.Core/Stryker.Core/Initialisation/Buildalyzer/IAnalyzerResultCSharpExtensions.cs
@@ -1,0 +1,41 @@
+using System;
+using Buildalyzer;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Stryker.Core.Options;
+
+namespace Stryker.Core.Initialisation.Buildalyzer
+{
+    public static class IAnalyzerResultCSharpExtensions
+    {
+        public static CSharpCompilationOptions GetCompilationOptions(this IAnalyzerResult analyzerResult)
+        {
+            var compilationOptions = new CSharpCompilationOptions(analyzerResult.GetOutputKind())
+                .WithNullableContextOptions(analyzerResult.GetNullableContextOptions())
+                .WithAllowUnsafe(analyzerResult.GetPropertyOrDefault("AllowUnsafeBlocks", true))
+                .WithAssemblyIdentityComparer(DesktopAssemblyIdentityComparer.Default)
+                .WithConcurrentBuild(true)
+                .WithModuleName(analyzerResult.GetAssemblyName())
+                .WithOverflowChecks(analyzerResult.GetPropertyOrDefault("CheckForOverflowUnderflow", false));
+
+            if (analyzerResult.IsSignedAssembly() && analyzerResult.GetAssemblyOriginatorKeyFile() is var keyFile && keyFile is not null)
+            {
+                compilationOptions = compilationOptions.WithCryptoKeyFile(keyFile)
+                    .WithStrongNameProvider(new DesktopStrongNameProvider());
+            }
+            return compilationOptions;
+        }
+
+        public static CSharpParseOptions GetParseOptions(this IAnalyzerResult analyzerResult, StrykerOptions options) => new CSharpParseOptions(options.LanguageVersion, DocumentationMode.None, preprocessorSymbols: analyzerResult.PreprocessorSymbols);
+
+        private static NullableContextOptions GetNullableContextOptions(this IAnalyzerResult analyzerResult)
+        {
+            if (!Enum.TryParse(analyzerResult.GetPropertyOrDefault("Nullable", "enable"), true, out NullableContextOptions nullableOptions))
+            {
+                nullableOptions = NullableContextOptions.Enable;
+            }
+
+            return nullableOptions;
+        }
+    }
+}
